requirements.txt
django
mixer
djangorestframework
djangorestframework-simplejwt
django-cors-headers